tests/backends/test_base.py
import pytest
from unittest.mock import MagicMock
from pyui_automation.backends.base import BaseBackend
from typing import Optional, List, Tuple, Any, Dict, Union
import numpy as np
from pathlib import Path


class ConcreteBackend(BaseBackend):
    """Concrete implementation of BaseBackend for testing"""

    def __init__(self) -> None:
        """Initialize ConcreteBackend with mock functions for testing"""
        self._mock_find_element = MagicMock()
        self._mock_find_elements = MagicMock()
        self._mock_get_active_window = MagicMock()
        self._mock_take_screenshot = MagicMock()
        self._mock_get_screen_size = MagicMock()
        self._mock_ocr = MagicMock()

    def find_element(self, by: str, value: str, timeout: float = 0) -> Optional[Any]:
        return self._mock_find_element(by, value)

    def find_elements(self, by: str, value: str) -> List[Any]:
        return self._mock_find_elements(by, value)

    def get_active_window(self) -> Optional[Any]:
        return self._mock_get_active_window()

    def take_screenshot(self, filepath: str) -> None:
        self._mock_take_screenshot(filepath)

    def get_screen_size(self) -> Tuple[int, int]:
        return self._mock_get_screen_size()

    def capture_screenshot(self) -> np.ndarray:
        return np.zeros((100, 100, 3), dtype=np.uint8)

    def capture_element_screenshot(self, element: Any) -> np.ndarray:
        return np.zeros((50, 50, 3), dtype=np.uint8)

    def click_mouse(self, x: int, y: int) -> None:
        pass

    def double_click_mouse(self, x: int, y: int) -> None:
        pass

    def right_click_mouse(self, x: int, y: int) -> None:
        pass

    def move_mouse(self, x: int, y: int) -> None:
        pass

    def press_key(self, key: str) -> None:
        pass

    def type_text(self, text: str) -> None:
        pass

    def get_window_title(self, window: Any) -> str:
        return "Test Window"

    def get_window_rect(self, window: Any) -> tuple[int, int, int, int]:
        return (0, 0, 800, 600)

    def maximize_window(self, window: Any) -> None:
        pass

    def minimize_window(self, window: Any) -> None:
        pass

    def restore_window(self, window: Any) -> None:
        pass

    def close_window(self, window: Any) -> None:
        pass

    def get_window_bounds(self, window: Any) -> Tuple[int, int, int, int]:
        return (0, 0, 800, 600)

    def check_accessibility(self, element: Any) -> Dict[str, Any]:
        return {}

    def get_element_attributes(self, element: Any) -> Dict[str, Any]:
        return {}

    def get_element_property(self, element: Any, property_name: str) -> Any:
        return None

    def set_element_property(self, element: Any, property_name: str, value: Any) -> None:
        pass

    def get_element_pattern(self, element: Any, pattern_name: str) -> Any:
        return None

    def invoke_element_pattern_method(self, pattern: Any, method_name: str, *args) -> Any:
        return None

    def get_element_rect(self, element: Any) -> tuple[int, int, int, int]:
        return (0, 0, 100, 100)

    def scroll_element(self, element: Any, direction: str, amount: float) -> None:
        pass

    def get_element_text(self, element: Any) -> str:
        return ""

    def set_element_text(self, element: Any, text: str) -> None:
        pass

    def get_element_value(self, element: Any) -> Any:
        return None

    def set_element_value(self, element: Any, value: Any) -> None:
        pass

    def get_element_state(self, element: Any) -> Dict[str, bool]:
        return {}

    def wait_for_element(self, by: str, value: str, timeout: float = 10) -> Optional[Any]:
        return None

    def wait_for_element_state(self, element: Any, state: str, timeout: float = 10) -> bool:
        return False

    def wait_for_element_property(self, element: Any, property_name: str, value: Any, timeout: float = 10) -> bool:
        return False

    def generate_accessibility_report(self, output_dir: Union[str, Path]) -> None:
        pass

    def get_application(self) -> Optional[Any]:
        return None

    def launch_application(self, path: str, *args, **kwargs) -> Any:
        return None

    def attach_to_application(self, process_id: int) -> Any:
        return None

    def close_application(self, application: Any) -> None:
        pass

    def get_window_handle(self, pid: Optional[int] = None) -> Optional[int]:
        return 0

    def release_key(self, key: str) -> None:
        pass

    def resize_window(self, window: Any, width: int, height: int) -> None:
        pass

    def send_keys(self, text: str) -> None:
        pass

    def set_window_position(self, window: Any, x: int, y: int) -> None:
        pass

    @property
    def application(self) -> Any:
        """Get the current application instance"""
        return MagicMock()

    @property
    def ocr(self) -> Any:
        """Get the OCR engine instance."""
        return self._mock_ocr

    def set_ocr_languages(self, languages: List[str]) -> None:
        """Set OCR languages for text recognition"""
        pass

    def move_mouse(self, x: int, y: int) -> None:
        """Move mouse cursor to absolute coordinates"""
        pass

    def click_mouse(self) -> None:
        """Click at current mouse position"""
        pass

    def double_click_mouse(self) -> None:
        """Double click at current mouse position"""
        pass

    def right_click_mouse(self) -> None:
        """Right click at current mouse position"""
        pass

    def mouse_down(self) -> None:
        """Press and hold primary mouse button"""
        pass

    def mouse_up(self) -> None:
        """Release primary mouse button"""
        pass

    def get_mouse_position(self) -> Tuple[int, int]:
        """Get current mouse cursor position"""
        return (0, 0)  # Return a default position for this incomplete implementation


class IncompleteBackend(BaseBackend):
    """Incomplete backend that doesn't implement all abstract methods"""

    def __init__(self):
        self._implemented_method = None
        self._mock_app = MagicMock()
        self._mock_window = MagicMock()
        self._mock_element = MagicMock()

    def find_element(self, by: str, value: str, timeout: float = 0) -> Optional[Any]:
        return None

    def find_elements(self, by: str, value: str) -> List[Any]:
        return []

    def get_active_window(self) -> Optional[Any]:
        return None

    def take_screenshot(self, filepath: str) -> None:
        pass

    def get_screen_size(self) -> Tuple[int, int]:
        return (800, 600)

    def capture_screenshot(self) -> np.ndarray:
        return np.zeros((100, 100, 3), dtype=np.uint8)

    def capture_element_screenshot(self, element: Any) -> np.ndarray:
        return np.zeros((50, 50, 3), dtype=np.uint8)

    def get_window_handle(self, pid: Optional[int] = None) -> Optional[Any]:
        return None

    def check_accessibility(self, element: Optional[Any] = None) -> Dict[str, Any]:
        return {}

    def click_mouse(self, x: int, y: int) -> None:
        pass

    def double_click_mouse(self, x: int, y: int) -> None:
        pass

    def right_click_mouse(self, x: int, y: int) -> None:
        pass

    def move_mouse(self, x: int, y: int) -> None:
        pass

    def press_key(self, key: str) -> None:
        pass

    def release_key(self, key: str) -> None:
        pass

    def send_keys(self, keys: str) -> None:
        pass

    def type_text(self, text: str) -> None:
        pass

    def get_window_title(self, window: Any) -> str:
        return ""

    def get_window_rect(self, window: Any) -> Tuple[int, int, int, int]:
        return (0, 0, 800, 600)

    def get_window_bounds(self, window: Any) -> Tuple[int, int, int, int]:
        return (0, 0, 800, 600)

    def set_window_position(self, window: Any, x: int, y: int) -> None:
        pass

    def resize_window(self, window: Any, width: int, height: int) -> None:
        pass

    def maximize_window(self, window: Any) -> None:
        pass

    def minimize_window(self, window: Any) -> None:
        pass

    def restore_window(self, window: Any) -> None:
        pass

    def close_window(self, window: Any) -> None:
        pass

    def get_element_attributes(self, element: Any) -> Dict[str, Any]:
        return {}

    def get_element_property(self, element: Any, property_name: str) -> Any:
        return None

    def set_element_property(self, element: Any, property_name: str, value: Any) -> None:
        pass

    def get_element_pattern(self, element: Any, pattern_name: str) -> Any:
        return None

    def invoke_element_pattern_method(self, pattern: Any, method_name: str, *args) -> Any:
        return None

    def get_element_rect(self, element: Any) -> tuple[int, int, int, int]:
        return (0, 0, 100, 100)

    def scroll_element(self, element: Any, direction: str, amount: float) -> None:
        pass

    def get_element_text(self, element: Any) -> str:
        return ""

    def set_element_text(self, element: Any, text: str) -> None:
        pass

    def get_element_value(self, element: Any) -> Any:
        return None

    def set_element_value(self, element: Any, value: Any) -> None:
        pass

    def get_element_state(self, element: Any) -> Dict[str, bool]:
        return {}

    def wait_for_element(self, by: str, value: str, timeout: float = 10) -> Optional[Any]:
        return None

    def wait_for_element_state(self, element: Any, state: str, timeout: float = 10) -> bool:
        return False

    def wait_for_element_property(self, element: Any, property_name: str, value: Any, timeout: float = 10) -> bool:
        return False

    def generate_accessibility_report(self, output_dir: Union[str, Path]) -> None:
        pass

    def get_application(self) -> Optional[Any]:
        return self._mock_app

    def launch_application(self, path: str, *args, **kwargs) -> Any:
        return self._mock_app

    def attach_to_application(self, process_id: int) -> Any:
        return self._mock_app

    def close_application(self, application: Any) -> None:
        pass

    def get_window_handle(self, pid: Optional[int] = None) -> Optional[int]:
        return 0

    @property
    def application(self) -> Any:
        """Get the current application instance"""
        return self._mock_app

    @property
    def ocr(self) -> Any:
        """Get the OCR engine instance."""
        return MagicMock()

    def set_ocr_languages(self, languages: List[str]) -> None:
        """Set OCR languages for text recognition"""
        pass

    def move_mouse(self, x: int, y: int) -> None:
        """Move mouse cursor to absolute coordinates"""
        pass

    def click_mouse(self) -> None:
        """Click at current mouse position"""
        pass

    def double_click_mouse(self) -> None:
        """Double click at current mouse position"""
        pass

    def right_click_mouse(self) -> None:
        """Right click at current mouse position"""
        pass

    def mouse_down(self) -> None:
        """Press and hold primary mouse button"""
        pass

    def mouse_up(self) -> None:
        """Release primary mouse button"""
        pass

    def get_mouse_position(self) -> Tuple[int, int]:
        """Get current mouse cursor position"""
        return (0, 0)  # Return a default position for this incomplete implementation


@pytest.fixture
def backend():
    """Create a concrete backend instance for testing"""
    return ConcreteBackend()


@pytest.fixture
def incomplete_backend():
    """Create an incomplete backend instance for testing"""
    return IncompleteBackend()


def test_find_element_by_id(backend):
    backend._mock_find_element.return_value = "test_element"
    result = backend.find_element("id", "test_id")
    assert result == "test_element"
    backend._mock_find_element.assert_called_once_with("id", "test_id")


def test_find_element_by_name(backend):
    backend._mock_find_element.return_value = "test_element"
    result = backend.find_element("name", "test_name")
    assert result == "test_element"
    backend._mock_find_element.assert_called_once_with("name", "test_name")


def test_find_element_by_class(backend):
    backend._mock_find_element.return_value = "test_element"
    result = backend.find_element("class", "test_class")
    assert result == "test_element"
    backend._mock_find_element.assert_called_once_with("class", "test_class")


def test_find_element_by_role(backend):
    backend._mock_find_element.return_value = "test_element"
    result = backend.find_element("role", "button")
    assert result == "test_element"
    backend._mock_find_element.assert_called_once_with("role", "button")


def test_find_element_by_xpath(backend):
    backend._mock_find_element.return_value = "test_element"
    result = backend.find_element("xpath", "//button[@id='test']")
    assert result == "test_element"
    backend._mock_find_element.assert_called_once_with("xpath", "//button[@id='test']")


def test_find_element_by_css(backend):
    backend._mock_find_element.return_value = "test_element"
    result = backend.find_element("css", "#test")
    assert result == "test_element"
    backend._mock_find_element.assert_called_once_with("css", "#test")


def test_find_element_by_text(backend):
    backend._mock_find_element.return_value = "test_element"
    result = backend.find_element("text", "Click me")
    assert result == "test_element"
    backend._mock_find_element.assert_called_once_with("text", "Click me")


def test_find_element_by_partial_text(backend):
    backend._mock_find_element.return_value = "test_element"
    result = backend.find_element("partial text", "Click")
    assert result == "test_element"
    backend._mock_find_element.assert_called_once_with("partial text", "Click")


def test_find_element_by_ocr(backend):
    backend._mock_find_element.return_value = "test_element"
    result = backend.find_element("ocr", "Click me")
    assert result == "test_element"
    backend._mock_find_element.assert_called_once_with("ocr", "Click me")


def test_find_element_by_image(backend):
    backend._mock_find_element.return_value = "test_element"
    result = backend.find_element("image", "button.png")
    assert result == "test_element"
    backend._mock_find_element.assert_called_once_with("image", "button.png")


def test_find_element_returns_none(backend):
    backend._mock_find_element.return_value = None
    result = backend.find_element("id", "nonexistent")
    assert result is None
    backend._mock_find_element.assert_called_once_with("id", "nonexistent")


def test_get_active_window(backend):
    backend._mock_get_active_window.return_value = "test_window"
    result = backend.get_active_window()
    assert result == "test_window"
    backend._mock_get_active_window.assert_called_once()


def test_take_screenshot(backend):
    backend._mock_take_screenshot.return_value = True
    result = backend.take_screenshot("test.png")
    assert result is None
    backend._mock_take_screenshot.assert_called_once_with("test.png")


def test_get_screen_size(backend):
    backend._mock_get_screen_size.return_value = (1920, 1080)
    result = backend.get_screen_size()
    assert result == (1920, 1080)
    backend._mock_get_screen_size.assert_called_once()


def test_find_elements(backend):
    expected = ["element1", "element2"]
    backend._mock_find_elements.return_value = expected
    result = backend.find_elements("class", "test")
    assert result == expected
    backend._mock_find_elements.assert_called_once_with("class", "test")


def test_find_elements_empty(backend):
    backend._mock_find_elements.return_value = []
    result = backend.find_elements("class", "nonexistent")
    assert result == []
    backend._mock_find_elements.assert_called_once_with("class", "nonexistent")


def test_abstract_find_element(incomplete_backend):
    with pytest.raises(NotImplementedError):
        incomplete_backend.find_element("id", "test")


def test_abstract_find_elements(incomplete_backend):
    with pytest.raises(NotImplementedError):
        incomplete_backend.find_elements("class", "test")


def test_abstract_get_active_window(incomplete_backend):
    with pytest.raises(NotImplementedError):
        incomplete_backend.get_active_window()


def test_abstract_take_screenshot(incomplete_backend):
    with pytest.raises(NotImplementedError):
        incomplete_backend.take_screenshot("test.png")


def test_abstract_get_screen_size(incomplete_backend):
    with pytest.raises(NotImplementedError):
        incomplete_backend.get_screen_size()
